Bugfix: let processes be stopped and resumed with signalfd(). Enhancement: let batch script exit successful if terminated by SIGUSR1     (pct:#133). Bugfix: send message task exit after closing of all I/O connections. Enhancement: set correct derived exit code on step timeout (pct:#20). Enhancement: improved log messages. Enhancement: correct error message for pam ssh logins if the prologue is     still running (pct:#171).
Update regression test suite:
 * Add huge-env tests
 * Bug fix for test multi-prog-5
 * Use cpu_bind=socket in mem_bind-local test
 * Revert "Add a new test mem_bind-local-2"
     This reverts commit f44a765adde19127f527ca5118a878b89879067a.
 * Add a new test mem_bind-local-2
 * Add test multi-prog-5

+++ include/pluginfrag.h
@@ -27,7 +27,7 @@ typedef struct {
     uint8_t uID;
     uint16_t msgNum;
     uint16_t msgCount;
-    uint16_t totalSize;
+    uint64_t totalSize;
 } PS_Frag_Msg_Header_t;
 
 typedef int Send_Msg_Func_t(void *);
